% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_guess.R
\name{make_guess}
\alias{make_guess}
\title{Makes a guess a the secret code}
\usage{
make_guess(combos)
}
\arguments{
\item{combos}{available combinations to guess at}
}
\value{
numeric vector representing a valid guess at secret code
}
\description{
Makes a guess a the secret code
}
\examples{
make_guess(combos = matrix(c(1,1,2,2,1,2,1,2), 4))

}
