% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_game.R
\name{print_game}
\alias{print_game}
\title{Print a particular game history to console}
\usage{
print_game(guess_history, response_history, color_scheme = NULL)
}
\arguments{
\item{guess_history}{list of numeric vectors representing guesses}

\item{response_history}{list of numeric vectors representing responses}

\item{color_scheme}{charactor with colors used for numbers from 0 to 9#'}
}
\description{
Print a particular game history to console
}
\examples{
print_game(guesses, responses)

}
