% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_guess.R
\name{rate_guess}
\alias{rate_guess}
\title{Rate a guess, given a set of available combinations}
\usage{
rate_guess(combos, guess)
}
\arguments{
\item{combos}{matrix of all available combinations}

\item{guess}{numeric vector with a particular guess}
}
\value{
integer. Maximum number of combinations we
can end up with after making this guess.
}
\description{
Rate a guess, given a set of available combinations
}
\examples{
rate_guess(combos = matrix(c(1,1,2,2,1,2,1,2), 4), guess = c(1, 1))
}
