% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate different strategies in variations of Mastermind}
\usage{
simulate(colors, pegs, strategy, n)
}
\arguments{
\item{colors}{integer from 1 to 9. Number of colors in game}

\item{pegs}{integer from 1 to 9. Number of pegs (slots) in game}

\item{strategy}{character, one of "Knuth", "Doable"}

\item{n}{integer, number of simulations to run}
}
\value{
vector with counts of the number of guesses used
}
\description{
Simulate different strategies in variations of Mastermind
}
\examples{
# Not run:
simulate(colors = 3, pegs = 2, strategy = "Knuth", n = 10)

}
