% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_guess.R
\name{make_guess}
\alias{make_guess}
\title{Makes a guess a the secret code}
\usage{
make_guess(all_combos, remaining_combos)
}
\arguments{
\item{all_combos}{tibble with all combinations to pick a guess from}

\item{remaining_combos}{tibble with available combinations to guess at}
}
\value{
numeric vector representing a valid guess at secret code
}
\description{
Makes a guess a the secret code
}
\examples{
co <- combos(2, 2)
make_guess(all_combos = co, remaining_combos = co)

}
