% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_guess.R
\name{rate_guess}
\alias{rate_guess}
\title{Rate a guess, given a set of available combinations}
\usage{
rate_guess(all_combos, remaining_combos, guess)
}
\arguments{
\item{all_combos}{matrix of all possible combinations, eg, colors^pegs}

\item{remaining_combos}{matrix of all remaining available combinations}

\item{guess}{numeric vector with a particular guess}
}
\value{
integer. Maximum number of combinations we
can end up with after making this guess.
}
\description{
Rate a guess, given a set of available combinations
}
\examples{
co <- combos(2, 2)
rate_guess(all_combos = co, remaining_combos = co, guess = c(1, 1))
}
