% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_guess.R
\name{rate_guess}
\alias{rate_guess}
\title{Rate a guess, given a set of available combinations}
\usage{
rate_guess(remaining_combos, guess, FUN = max, ...)
}
\arguments{
\item{remaining_combos}{matrix of all remaining available combinations}

\item{guess}{numeric vector with a particular guess}

\item{FUN}{function used to summarise the number of remaining combinations
after a guess followed by a response}

\item{...}{parameters to be passed on to FUN if any}
}
\value{
Summary of the number of combinations we can end up with
    after making this guess. Default is max(). So, after guessing
    at a code like 1234, any response will leave us with at
    \emph{most} this number of remaining possible combinations.
}
\description{
Rate a guess, given a set of available combinations
}
\examples{
rate_guess(remaining_combos = combos(2, 2), guess = c(1, 1))
}
