% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_combos.R
\name{reduce_combos}
\alias{reduce_combos}
\title{Find subset of available combinations that are consistent with
a given guess and response}
\usage{
reduce_combos(combos, guess, response)
}
\arguments{
\item{combos}{matrix of all available combinations}

\item{guess}{numeric vector with a particular guess}

\item{response}{named numeric vector of length 2 with response from that guess}
}
\value{
matrix, subset of the original
}
\description{
Find subset of available combinations that are consistent with
a given guess and response
}
\examples{
reduce_combos(combos, guess, response)

}
